+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 2.0.1
 - Shift clicking upgrades doesn't put the upgrades into the filter slots anymore
+- Improved grid sorting performance
 
 ### 2.0.0
 - Added the Wireless Crafting Grid